Move some more files around, bump up to C++14 for GPU
