+++ runtime/include/gpu/amd/rocm-utils.h
